moved tests to test folder to fix deployment errors
