+++ src/__tests__/header.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen, fireEvent } from "@testing-library/react";
-import Header from "./header";
+import Header from "../components/header";
 
 describe("Header Component", () => {
   it("renders the header with all elements", () => {
